backend/requirements.txt
fastapi
uvicorn
python-multipart
qdrant-client
sentence-transformers
openai
python-dotenv
Pillow
numpy
requests
groq